api/requirements.txt
numpy
pandas
tensorflow
matplotlib
pillow
Flask
flask-cors
opencv-python
waitress
build
scikit-learn